requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
loguru==0.7.2
httpx==0.27.2
beautifulsoup4
pandas
numpy
requests
python-multipart
google-generativeai
